pyproject.toml
[tool.poetry]
name = "iqnews"
version = "0.9.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aio-pika = "9.5.3"
aiofiles = "24.1.0"
aiogram = "3.15.0"
aiohappyeyeballs = "2.4.4"
aiohttp = "3.10.11"
aiolimiter = "1.2.1"
aiormq = "6.8.1"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.7.0"
async-timeout = "5.0.1"
asyncpg = "0.30.0"
attrs = "24.2.0"
beautifulsoup4 = "4.12.3"
black = "24.10.0"
blessed = "1.20.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
distro = "1.9.0"
editor = "1.6.6"
exceptiongroup = "1.2.2"
feedparser = "6.0.11"
frozenlist = "1.5.0"
greenlet = "3.1.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
inquirer = "3.4.0"
isort = "5.13.2"
jiter = "0.8.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain = "0.3.11"
langchain-core = "0.3.24"
langchain-openai = "0.2.12"
langchain-text-splitters = "0.3.2"
langchain-together = "0.2.0"
langsmith = "0.2.3"
magic-filter = "1.0.12"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
numpy = "2.2.0"
openai = "1.57.3"
orjson = "3.10.12"
packaging = "24.2"
pamqp = "3.3.0"
pathspec = "0.12.1"
pika = "1.3.2"
platformdirs = "4.3.6"
playwright = "1.49.1"
propcache = "0.2.1"
psycopg2 = "2.9.10"
psycopg2-binary = "2.9.10"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pyee = "12.0.0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pyuseragents = "1.0.5"
pyyaml = "6.0.2"
readchar = "4.2.1"
redis = "5.2.1"
regex = "2024.11.6"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
ruff = "0.8.3"
runs = "1.2.2"
safeio = "1.2"
setuptools = "75.1.0"
sgmllib3k = "1.0.0"
six = "1.17.0"
sniffio = "1.3.1"
soupsieve = "2.6"
sqlalchemy = "2.0.36"
tenacity = "9.0.0"
tiktoken = "0.8.0"
tqdm = "4.67.1"
translatepy = "2.3"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
wcwidth = "0.2.13"
wheel = "0.44.0"
xmod = "1.8.1"
yarl = "1.18.3"
aioredis = "2.0.1"
prometheus-client = "^0.21.1"
aioresponses = "^0.7.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

[tool.ruff]
# Enable rules from various tools
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear (common pitfalls)
    "S",    # flake8-bandit (security issues)
    "PL",   # pylint
]
exclude = [
    "tests",
    "services/tg_bot/texts.py",
    "*/prompts.py",
    "*/config.py",
]
ignore = [
    "E501", # Stupid line length
    "PLR0913", # Important for logging
    "S110",
    "PLR1704",
    "PLR0911",
]
